third_party/blink/web_tests/external/wpt/cobalt/h5vcc-system/h5vcc-system.https.sub.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test H5vccSystem web API.</title>
  <script src='/resources/testharness.js'></script>
  <script src='/resources/testharnessreport.js'></script>
</head>
<body>
  <script>
    test(() => {
        assert_equals(window.h5vcc.system.advertisingId, "fake advertisingId");
    }, "H5vccSystem has the right property value");
  </script>
</body>
</html>
